biomapas_repo/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
scikit-learn==1.2.1
spacy==3.5.3
numpy==1.23.5
